Configured the apps and users pages for admin with delete and freeze logic

+++ server/controllers/AdminController.js
@@ -1,4 +1,5 @@
 const AppModel = require("../models/App");
+const InvoiceModel = require("../models/Invocie");
 const IssueModel = require("../models/Issue");
 const UserModel = require("../models/User");
 const { getActivityToday } = require("../services/activityService");
@@ -31,10 +32,10 @@ const getOverview = async (_, res) => {
 };
 
 const deleteUser = async (req, res) => {
-  const { user, deleteCommand } = req.body;
+  const { username, deleteCommand } = req.body;
 
   try {
-    const isDeleteConfirmed = deleteCommand.toLowerCase() === `delete ${user.toLowerCase()}`;
+    const isDeleteConfirmed = deleteCommand.toLowerCase() === `delete ${username.toLowerCase()}`;
 
     if (!isDeleteConfirmed) {
       return res
@@ -42,7 +43,7 @@ const deleteUser = async (req, res) => {
         .json({ message: "Deletion failed due to the absence of confirmation", success: false });
     }
 
-    const userData = await UserModel.findOne({ name: user });
+    const userData = await UserModel.findOne({ name: username });
 
     if (!userData) {
       return res.status(404).json({ message: "User not found", success: false });
@@ -54,9 +55,11 @@ const deleteUser = async (req, res) => {
       return res.status(400).json({ message: "Cannot delete admin account", success: false });
     }
 
-    await UserModel.deleteOne({ name: user });
+    await UserModel.deleteOne({ name: username });
 
-    res.status(204).json({ message: "User deleted successfully", success: true });
+    const users = await getUsers();
+
+    res.status(200).json({ message: "User deleted successfully", success: true, users });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Internal server error", success: false });
@@ -64,10 +67,10 @@ const deleteUser = async (req, res) => {
 };
 
 const freezeAccount = async (req, res) => {
-  const { user } = req.body;
+  const { username } = req.body;
 
   try {
-    const userData = await UserModel.findOne({ name: user });
+    const userData = await UserModel.findOne({ name: username });
 
     if (!userData) {
       return res.status(404).json({ message: "User not found", success: false });
@@ -85,7 +88,29 @@ const freezeAccount = async (req, res) => {
     userData.isFrozen = !isFrozen;
     await userData.save();
 
-    res.status(200).json({ message, status: true });
+    let newInvoice;
+
+    if (isFrozen) {
+      newInvoice = new InvoiceModel({
+        subject: "Account Access Notification",
+        to: username,
+        from: "Authecho",
+        text: `Dear ${username},\n\nWe are pleased to inform you that you have successfully regained access to your account. The issue previously reported regarding your account has been thoroughly investigated and resolved. We appreciate your patience during this process and are committed to ensuring the security of your account.\n\nBest regards,\nThe Authecho Team`,
+      });
+    } else {
+      newInvoice = new InvoiceModel({
+        subject: "Account Access Notification",
+        to: username,
+        from: "Authecho",
+        text: `Dear ${username},\n\nWe wish to formally inform you that your account has been temporarily frozen due to suspicious activity that may indicate unauthorized access. As we conduct a thorough investigation into this matter, you will regrettably be unable to access your account during this period. We appreciate your understanding and patience as we work to resolve this issue promptly.\n\nBest regards,\nThe Authecho Team`,
+      });
+    }
+
+    await newInvoice.save();
+
+    const users = await getUsers();
+
+    res.status(200).json({ message, success: true, users });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Internal server error", success: false });
@@ -104,9 +129,26 @@ const deleteApp = async (req, res) => {
         .json({ message: "Deletion failed due to the absence of confirmation", success: false });
     }
 
+    const appData = await AppModel.findOne({ name: app });
+
+    if (!appData) {
+      return res.status(404).json({ message: "App not found", success: false });
+    }
+
     await AppModel.deleteOne({ name: app });
 
-    res.status(204).json({ message: "App deleted successfully", success: true });
+    const newInvoice = new InvoiceModel({
+      subject: "Application Deletion Notification",
+      to: appData.creator,
+      from: "Authecho",
+      text: `Dear ${appData.creator},\n\nWe regret to inform you that the application named "${app}" has been permanently removed from our database. This action may have been taken due to consistently low activity associated with your application; however, it is more likely a result of the usage of access controls that do not align with our policies. We encourage you to review our guidelines to ensure compliance for future applications.\n\nThank you for your understanding.\n\nBest regards,\nThe Authecho Team`,
+    });
+
+    await newInvoice.save();
+
+    const apps = await getApps();
+
+    res.status(200).json({ message: "App deleted successfully", success: true, apps });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Internal server error", success: false });
@@ -129,7 +171,29 @@ const freezeApp = async (req, res) => {
     appData.isFrozen = !isFrozen;
     await appData.save();
 
-    res.status(200).json({ message, status: true });
+    let newInvoice;
+
+    if (isFrozen) {
+      newInvoice = new InvoiceModel({
+        subject: "Application Access Notification",
+        to: appData.creator,
+        from: "Authecho",
+        text: `Dear ${appData.creator},\n\nWe are pleased to inform you that you have successfully regained access to your application, "${app}". This restoration follows a thorough investigation into the reasons behind the temporary suspension of access. We appreciate your patience during this process and are committed to ensuring the security of your application.\n\nBest regards,\nThe Authecho Team`,
+      });
+    } else {
+      newInvoice = new InvoiceModel({
+        subject: "Application Access Notification",
+        to: appData.creator,
+        from: "Authecho",
+        text: `Dear ${appData.creator},\n\nWe regret to inform you that access to an application connected to Authecho has been temporarily suspended. This action may be due to suspicious activity related to your application or a vulnerability identified in its security. We are actively investigating this matter and recommend reviewing your application for any potential issues.\n\nThank you for your understanding.\n\nBest regards,\nThe Authecho Team`,
+      });
+    }
+
+    await newInvoice.save();
+
+    const apps = await getApps();
+
+    res.status(200).json({ message, status: true, apps });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Internal server error", success: false });
@@ -142,7 +206,9 @@ const deleteIssue = async (req, res) => {
   try {
     await IssueModel.deleteOne({ _id: issueID });
 
-    res.status(204).json({ message: "Issue deleted successfully", success: true });
+    const issues = await getIssues();
+
+    res.status(204).json({ message: "Issue deleted successfully", success: true, issues });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Internal server error", success: false });
@@ -165,7 +231,9 @@ const resolveIssue = async (req, res) => {
     issueData.isResolved = !isResolved;
     await issueData.save();
 
-    res.status(200).json({ message, status: true });
+    const issues = await getIssues();
+
+    res.status(200).json({ message, status: true, issues });
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: "Internal server error", success: false });
